test(e2e): skip watch tests when the live API withholds streams
The watch page guard checked the video title before the error message and
short-circuited to "loaded". An IP block still hydrates the title,
description and recommendations, so the guard passed and the bare player
assertion right after it failed instead, before waitForPlaybackOrSkip
could recognise the block.

Move the player into the settle condition, let the error win over a
loaded title, and share the guard between the two specs that duplicated
it. The timeout covers the reload-once IP block recovery.

Also skip the Shorts thumbnail test when the API keeps answering with an
empty Shorts tab, and follow the renamed stream extraction method label
in the settings migration test.
