+++ e2e/tests/network/watch.spec.mjs
@@ -1,6 +1,6 @@
 import { sel } from '../../helpers/app.mjs'
 import { test, expect, setPlayerFullscreen } from '../../helpers/innertube.mjs'
-import { activeTab, findWatchComponent, waitForPlaybackOrSkip } from '../../helpers/player.mjs'
+import { activeTab, findWatchComponent, openVideoOrSkip, waitForPlaybackOrSkip } from '../../helpers/player.mjs'
 
 // "Me at the zoo" - the oldest video on YouTube, short and stable.
 const VIDEO_URL = 'https://www.youtube.com/watch?v=jNQXAC9IVRw'
@@ -45,34 +45,8 @@ function longTranscript() {
   )).join('\n\n')}\n`
 }
 
-/**
- * Opens the watch page, or skips when the live API refuses to serve it (bot
- * checks on CI runners and VPNs), which leaves the page shell without any
- * video data instead of producing an error message.
- */
-async function openVideo(page, video = { id: 'jNQXAC9IVRw', title: 'Me at the zoo', url: VIDEO_URL }) {
-  await page.locator(sel.searchInput).fill(video.url)
-  await page.locator(sel.searchInput).press('Enter')
-  await expect(page).toHaveURL(new RegExp(`#\\/watch\\/${video.id}`))
-
-  const title = page.locator(`${activeTab} .videoTitle`)
-  const errorMessage = page.locator(`${activeTab} .errorMessage`)
-  let state = 'waiting'
-  const settled = await expect
-    .poll(async () => {
-      const titleText = await title.textContent().catch(() => '') ?? ''
-      if (titleText.includes(video.title)) {
-        state = 'loaded'
-      } else if (await errorMessage.isVisible().catch(() => false)) {
-        state = (await errorMessage.textContent())?.trim() ?? 'unavailable'
-      }
-      return state === 'waiting' ? 'waiting' : 'done'
-    }, { timeout: 30_000, message: 'waiting for the watch page to load' })
-    .toBe('done')
-    .then(() => true, () => false)
-
-  test.skip(!settled || state !== 'loaded', `watch page unavailable from the live API: ${state}`)
-  await expect(page.locator(`${activeTab} .ftVideoPlayer`)).toBeVisible({ timeout: 30_000 })
+function openVideo(page, video = { id: 'jNQXAC9IVRw', title: 'Me at the zoo', url: VIDEO_URL }) {
+  return openVideoOrSkip(test, page, video)
 }
 
 async function openCaptionedVideoOrSkip(page) {
